+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 
 ### Bug Fixes
 
-- Fix inconsistant order of participants when sorted by the raised hand. ([#1630](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1630))
+- Fix thumbnail row update on participant leave in speaker view ([#1468](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1468))
 
+- Fix inconsistant order of participants when sorted by the raised hand. ([#1630](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1630))
 
 ## 1.10.0